@@ -1,83 +1,79 @@
 {
-    "name": "openapi-axios",
-    "version": "0.25.0",
-    "description": "OpenAPI(2.0/3.0/3.1) Schema → Type-safe Axios",
-    "scripts": {
-        "prepare": "husky install",
-        "build": "NODE_ENV=production vite build",
-        "lint:js": "eslint --ext=.cjs,.mjs,.ts,.tsx src test",
-        "lint:js:fix": "eslint --ext=.cjs,.mjs,.ts,.tsx src test --fix",
-        "lint:type": " tsc --project tsconfig.json",
-        "lint": "npm run lint:js && npm run lint:type",
-        "lint:fix": "npm run lint:js:fix && npm run lint:type",
-        "test": "vitest run",
-        "test:coverage": "vitest run --coverage"
+  "name": "openapi-axios",
+  "type": "commonjs",
+  "version": "0.25.0",
+  "description": "OpenAPI(2.0/3.0/3.1) Schema → Type-safe Axios",
+  "author": "云淡然 <hi@ydr.me> (https://ydr.me)",
+  "license": "MIT",
+  "homepage": "https://github.com/FrontEndDev-org/openapi-axios",
+  "repository": "https://github.com/FrontEndDev-org/openapi-axios",
+  "keywords": [
+    "cloudcome",
+    "ydr.me",
+    "OpenAPI",
+    "swagger",
+    "TypeScript",
+    "OAS",
+    "axios"
+  ],
+  "sideEffects": false,
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "import": "./dist/index.mjs",
+      "require": "./dist/index.cjs"
     },
-    "engines": {
-        "node": ">=14.21"
-    },
-    "sideEffects": false,
-    "type": "commonjs",
-    "main": "./dist/index.cjs",
-    "module": "./dist/index.mjs",
-    "exports": {
-        ".": {
-            "types": "./dist/index.d.ts",
-            "import": "./dist/index.mjs",
-            "require": "./dist/index.cjs"
-        },
-        "./package.json": "./package.json"
-    },
-    "types": "./dist/index.d.ts",
-    "files": [
-        "dist"
-    ],
-    "bin": {
-        "openapi-axios": "bin/index.cjs"
-    },
-    "keywords": [
-        "cloudcome",
-        "ydr.me",
-        "OpenAPI",
-        "swagger",
-        "TypeScript",
-        "OAS",
-        "axios"
-    ],
-    "author": "云淡然 <hi@ydr.me> (https://ydr.me)",
-    "maintainers": [
-        "阿森 <zhenxs2018@foxmail.com>"
-    ],
-    "homepage": "https://github.com/FrontEndDev-org/openapi-axios",
-    "repository": "https://github.com/FrontEndDev-org/openapi-axios",
-    "license": "MIT",
-    "dependencies": {
-        "@redocly/openapi-core": "^1.26.1",
-        "chalk": "^4.1.2",
-        "commander": "^12.1.0",
-        "prettier": "^3.4.2",
-        "strict-event-emitter": "^0.5.1",
-        "try-flatten": "^1.3.2"
-    },
-    "devDependencies": {
-        "@commitlint/cli": "^18.6.1",
-        "@commitlint/config-conventional": "^18.6.3",
-        "@commitlint/types": "^18.6.1",
-        "@types/node": "20",
-        "@types/prettier": "^3.0.0",
-        "@typescript-eslint/eslint-plugin": "^7.18.0",
-        "@typescript-eslint/parser": "^7.18.0",
-        "@vitest/coverage-v8": "^1.6.0",
-        "eslint": "^8.57.1",
-        "eslint-config-prettier": "^9.1.0",
-        "eslint-define-config": "^2.1.0",
-        "eslint-plugin-prettier": "^5.2.1",
-        "husky": "^8.0.3",
-        "lint-staged": "^15.3.0",
-        "typescript": "^5.7.2",
-        "vite": "^5.4.11",
-        "vite-plugin-dts": "^3.9.1",
-        "vite-plugin-externalize-deps": "^0.8.0",
-        "vitest": "^1.6.0"
-    }
+    "./package.json": "./package.json"
+  },
+  "main": "./dist/index.cjs",
+  "module": "./dist/index.mjs",
+  "types": "./dist/index.d.ts",
+  "bin": {
+    "openapi-axios": "bin/index.cjs"
+  },
+  "files": [
+    "dist"
+  ],
+  "engines": {
+    "node": ">=14.21"
+  },
+  "scripts": {
+    "prepare": "husky",
+    "build": "NODE_ENV=production vite build",
+    "lint:js": "eslint",
+    "lint:js:fix": "eslint --fix",
+    "lint:type": " tsc --project tsconfig.json",
+    "lint": "npm run lint:js && npm run lint:type",
+    "lint:fix": "npm run lint:js:fix && npm run lint:type",
+    "test": "vitest run",
+    "test:coverage": "vitest run --coverage"
+  },
+  "maintainers": [
+    "阿森 <zhenxs2018@foxmail.com>"
+  ],
+  "dependencies": {
+    "@redocly/openapi-core": "^1.26.1",
+    "chalk": "^4.1.2",
+    "commander": "^13.0.0",
+    "prettier": "^3.4.2",
+    "strict-event-emitter": "^0.5.1",
+    "try-flatten": "^1.3.2"
+  },
+  "devDependencies": {
+    "@antfu/eslint-config": "^3.12.1",
+    "@commitlint/cli": "^19.6.1",
+    "@commitlint/config-conventional": "^19.6.0",
+    "@commitlint/types": "^19.5.0",
+    "@types/node": "20",
+    "@vitest/coverage-v8": "^1.6.0",
+    "eslint": "^9.17.0",
+    "husky": "^9.1.7",
+    "lint-staged": "^15.3.0",
+    "taze": "^0.18.0",
+    "typescript": "^5.7.2",
+    "vite": "^5.4.11",
+    "vite-plugin-dts": "^3.9.1",
+    "vite-plugin-externalize-deps": "^0.8.0",
+    "vitest": "^1.6.0"
+  }
 }
