@@ -1,39 +1,41 @@
+import fs from 'node:fs';
+import path from 'node:path';
+import process from 'node:process';
 import { Command } from 'commander';
-import fs from 'fs';
-import path from 'path';
+import { generate, resolveConfigFile } from './command';
 import { pkgName, pkgVersion } from './const';
-import { resolveConfigFile, generate } from './command';
 
 export function createCLI() {
-    const program = new Command();
+  const program = new Command();
 
-    program
-        .name(pkgName)
-        .version(pkgVersion)
-        .description(process.env.PKG_DESCRIPTION)
-        .action((options, command) => {
-            if (command.args.length === 0) return generate();
+  program
+    .name(pkgName)
+    .version(pkgVersion)
+    .description(process.env.PKG_DESCRIPTION)
+    .action((options, command) => {
+      if (command.args.length === 0)
+        return generate();
 
-            program.help();
-        });
+      program.help();
+    });
 
-    program
-        .command('init')
-        .description('初始化配置文件')
-        .action(async () => {
-            const configFile = resolveConfigFile(process.cwd());
+  program
+    .command('init')
+    .description('初始化配置文件')
+    .action(async () => {
+      const configFile = resolveConfigFile(process.cwd());
 
-            if (configFile) {
-                console.log(`配置文件已存在`, path.relative(process.cwd(), configFile));
-                return;
-            }
+      if (configFile) {
+        console.log(`配置文件已存在`, path.relative(process.cwd(), configFile));
+        return;
+      }
 
-            const configFilename = 'openapi.config.cjs';
-            const configFilePath = path.join(process.cwd(), configFilename);
+      const configFilename = 'openapi.config.cjs';
+      const configFilePath = path.join(process.cwd(), configFilename);
 
-            fs.writeFileSync(
-                configFilePath,
-                `
+      fs.writeFileSync(
+        configFilePath,
+        `${`
 const { defineConfig } = require('openapi-axios');
 
 /**
@@ -44,10 +46,10 @@ module.exports = defineConfig({
     modules: {
         'petStore3': 'https://petstore31.swagger.io/api/v31/openapi.json'
     },
-});`.trim() + '\n',
-            );
-            console.log('生成配置文件', configFilename);
-        });
+});`.trim()}\n`,
+      );
+      console.log('生成配置文件', configFilename);
+    });
 
-    program.parse();
+  program.parse();
 }
