@@ -1,55 +1,58 @@
-import fs from 'fs';
-import path from 'path';
-import process from 'process';
-import type { OpenAPIAll, OpenAPILatest } from '../types/openapi';
-import { isString } from '../utils/type-is';
+import type { OpenAPIAll } from '../types/openapi';
+import fs from 'node:fs';
+import path from 'node:path';
+import process from 'node:process';
+import { bundleFromString, createConfig } from '@redocly/openapi-core';
 import { migrate } from '../migrations';
-import { bundle, bundleFromString, createConfig } from '@redocly/openapi-core';
+import { isString } from '../utils/type-is';
 
 type AcceptDocument = OpenAPIAll.Document | string;
 
 export class Reader {
-    cwd = process.cwd();
+  cwd = process.cwd();
 
-    async read(document: AcceptDocument) {
-        const source = await this._read(document);
-        const config = await createConfig({});
-        const { problems, bundle } = await bundleFromString({ config, source });
+  async read(document: AcceptDocument) {
+    const source = await this._read(document);
+    const config = await createConfig({});
+    const { problems, bundle } = await bundleFromString({ config, source });
 
-        if (problems.length) {
-            console.warn(`发现了 ${problems.length} 处错误，请检查文档，可能会出现非预期错误`);
-            problems.forEach((p) => {
-                console.warn(p.message);
-            });
-        }
-
-        return migrate(bundle.parsed);
+    if (problems.length) {
+      console.warn(`发现了 ${problems.length} 处错误，请检查文档，可能会出现非预期错误`);
+      problems.forEach((p) => {
+        console.warn(p.message);
+      });
     }
 
-    private async _read(document: AcceptDocument): Promise<string> {
-        if (isString(document)) {
-            if (/^https?:/i.test(document)) {
-                return await this.readRemote(document);
-            } else {
-                return this.readLocal(document);
-            }
-        } else {
-            return JSON.stringify(document);
-        }
-    }
+    return migrate(bundle.parsed);
+  }
 
-    protected readLocal(file: string) {
-        return fs.readFileSync(path.resolve(this.cwd, file), 'utf8');
+  private async _read(document: AcceptDocument): Promise<string> {
+    if (isString(document)) {
+      if (/^https?:/i.test(document)) {
+        return await this.readRemote(document);
+      }
+      else {
+        return this.readLocal(document);
+      }
     }
-
-    protected async readRemote(url: string) {
-        const resp = await fetch(url);
-        if (resp.ok) return await resp.text();
-
-        throw new Error(`${resp.status} ${resp.statusText}`);
+    else {
+      return JSON.stringify(document);
     }
+  }
 
-    protected readObject(document: OpenAPIAll.Document) {
-        return document;
-    }
+  protected readLocal(file: string) {
+    return fs.readFileSync(path.resolve(this.cwd, file), 'utf8');
+  }
+
+  protected async readRemote(url: string) {
+    const resp = await fetch(url);
+    if (resp.ok)
+      return await resp.text();
+
+    throw new Error(`${resp.status} ${resp.statusText}`);
+  }
+
+  protected readObject(document: OpenAPIAll.Document) {
+    return document;
+  }
 }
