@@ -1,67 +1,66 @@
-import { INTERNAL_TYPES, INTERNAL_VARS, KEYWORD_VARS } from './const';
 import { fixVarName, nextUniqueName } from '../utils/string';
-import { isString } from '../utils/type-is';
+import { INTERNAL_TYPES, INTERNAL_VARS, KEYWORD_VARS } from './const';
 
 export class Named {
-    varNameCountMap = new Map<string /*var*/, number /*count*/>();
-    typeNameCountMap = new Map<string /*type*/, number /*count*/>();
-    refIdTypeMap = new Map<string /*refId*/, string /*refType*/>();
+  varNameCountMap = new Map<string /* var */, number /* count */>();
+  typeNameCountMap = new Map<string /* type */, number /* count */>();
+  refIdTypeMap = new Map<string /* refId */, string /* refType */>();
 
-    constructor() {
-        KEYWORD_VARS.forEach(this.internalVarName.bind(this));
-        INTERNAL_VARS.forEach(this.internalVarName.bind(this));
-        INTERNAL_TYPES.forEach(this.internalTypeName.bind(this));
-    }
+  constructor() {
+    KEYWORD_VARS.forEach(this.internalVarName.bind(this));
+    INTERNAL_VARS.forEach(this.internalVarName.bind(this));
+    INTERNAL_TYPES.forEach(this.internalTypeName.bind(this));
+  }
 
-    /**
-     * 注册内部变量
-     * @param {string} varName
-     */
-    internalVarName(varName: string) {
-        this.varNameCountMap.set(varName, 1);
-    }
+  /**
+   * 注册内部变量
+   * @param {string} varName
+   */
+  internalVarName(varName: string) {
+    this.varNameCountMap.set(varName, 1);
+  }
 
-    /**
-     * 注册内部类型
-     * @param {string} typeName
-     */
-    internalTypeName(typeName: string) {
-        this.typeNameCountMap.set(typeName, 1);
-    }
+  /**
+   * 注册内部类型
+   * @param {string} typeName
+   */
+  internalTypeName(typeName: string) {
+    this.typeNameCountMap.set(typeName, 1);
+  }
 
-    nextVarName(name: string) {
-        return nextUniqueName(fixVarName(name), this.varNameCountMap);
-    }
+  nextVarName(name: string) {
+    return nextUniqueName(fixVarName(name), this.varNameCountMap);
+  }
 
-    nextOperationId(method: string, url: string, operationId?: string) {
-        operationId =
-            operationId ||
-            fixVarName(
-                [
-                    method,
-                    url
-                        .replace(/\{.*?}/g, '')
-                        .split('/')
-                        .filter(Boolean),
-                ].join('_'),
+  nextOperationId(method: string, url: string, operationId?: string) {
+    operationId
+            = operationId
+            || fixVarName(
+              [
+                method,
+                url
+                  .replace(/\{.*?\}/g, '')
+                  .split('/')
+                  .filter(Boolean),
+              ].join('_'),
             );
-        return nextUniqueName(operationId, this.varNameCountMap);
-    }
+    return nextUniqueName(operationId, this.varNameCountMap);
+  }
 
-    nextTypeName(typeName: string) {
-        const fixedTypeName = fixVarName(typeName, true, 'Type');
-        return nextUniqueName(fixedTypeName, this.typeNameCountMap);
-    }
+  nextTypeName(typeName: string) {
+    const fixedTypeName = fixVarName(typeName, true, 'Type');
+    return nextUniqueName(fixedTypeName, this.typeNameCountMap);
+  }
 
-    nextRefType(refType: string, refId: string) {
-        const uniqueTypeName = this.nextTypeName(refType);
+  nextRefType(refType: string, refId: string) {
+    const uniqueTypeName = this.nextTypeName(refType);
 
-        this.refIdTypeMap.set(refId, uniqueTypeName);
+    this.refIdTypeMap.set(refId, uniqueTypeName);
 
-        return uniqueTypeName;
-    }
+    return uniqueTypeName;
+  }
 
-    getRefType(ref: string) {
-        return this.refIdTypeMap.get(ref) || '';
-    }
+  getRefType(ref: string) {
+    return this.refIdTypeMap.get(ref) || '';
+  }
 }
