@@ -1,95 +1,93 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-/* eslint-disable @typescript-eslint/ban-types */
-/* eslint-disable @typescript-eslint/no-namespace */
+/* eslint-disable ts/no-empty-object-type */
 
-import * as OpenAPIV3 from './openapi-3_0';
+import type * as OpenAPIV3 from './openapi-3_0';
 
 export type Modify<T, R> = Omit<T, keyof R> & R;
-export type PathsWebhooksComponents<T extends {} = {}> = {
-    paths: PathsObject<T>;
-    webhooks: Record<string, PathItemObject | ReferenceObject>;
-    components: ComponentsObject;
-};
+export interface PathsWebhooksComponents<T extends {} = {}> {
+  paths: PathsObject<T>;
+  webhooks: Record<string, PathItemObject | ReferenceObject>;
+  components: ComponentsObject;
+}
 export type Document<T extends {} = {}> = Modify<
-    Omit<OpenAPIV3.Document<T>, 'paths' | 'components'>,
-    {
-        info: InfoObject;
-        jsonSchemaDialect?: string;
-        servers?: ServerObject[];
-    } & (
+  Omit<OpenAPIV3.Document<T>, 'paths' | 'components'>,
+  {
+    info: InfoObject;
+    jsonSchemaDialect?: string;
+    servers?: ServerObject[];
+  } & (
         | (Pick<PathsWebhooksComponents<T>, 'paths'> & Omit<Partial<PathsWebhooksComponents<T>>, 'paths'>)
         | (Pick<PathsWebhooksComponents<T>, 'webhooks'> & Omit<Partial<PathsWebhooksComponents<T>>, 'webhooks'>)
         | (Pick<PathsWebhooksComponents<T>, 'components'> & Omit<Partial<PathsWebhooksComponents<T>>, 'components'>)
     )
 >;
 export type InfoObject = Modify<
-    OpenAPIV3.InfoObject,
-    {
-        summary?: string;
-        license?: LicenseObject;
-    }
+  OpenAPIV3.InfoObject,
+  {
+    summary?: string;
+    license?: LicenseObject;
+  }
 >;
 export type ContactObject = OpenAPIV3.ContactObject;
 export type LicenseObject = Modify<
-    OpenAPIV3.LicenseObject,
-    {
-        identifier?: string;
-    }
+  OpenAPIV3.LicenseObject,
+  {
+    identifier?: string;
+  }
 >;
 export type ServerObject = Modify<
-    OpenAPIV3.ServerObject,
-    {
-        url: string;
-        description?: string;
-        variables?: Record<string, ServerVariableObject>;
-    }
+  OpenAPIV3.ServerObject,
+  {
+    url: string;
+    description?: string;
+    variables?: Record<string, ServerVariableObject>;
+  }
 >;
 export type ServerVariableObject = Modify<
-    OpenAPIV3.ServerVariableObject,
-    {
-        enum?: [string, ...string[]];
-    }
+  OpenAPIV3.ServerVariableObject,
+  {
+    enum?: [string, ...string[]];
+  }
 >;
 export type PathsObject<T extends {} = {}, P extends {} = {}> = Record<string, (PathItemObject<T> & P) | ReferenceObject>;
 export type HttpMethods = OpenAPIV3.HttpMethods;
 export type PathItemObject<T extends {} = {}> = {
-    $id?: string;
-    summary?: string;
-    description?: string;
-    servers?: ServerObject[];
-    parameters?: (ReferenceObject | ParameterObject)[];
+  $id?: string;
+  summary?: string;
+  description?: string;
+  servers?: ServerObject[];
+  parameters?: (ReferenceObject | ParameterObject)[];
 } & {
-    [method in HttpMethods]?: OperationObject<T>;
+  [method in HttpMethods]?: OperationObject<T>;
 };
 export type OperationObject<T extends {} = {}> = Modify<
-    OpenAPIV3.OperationObject<T>,
-    {
-        parameters?: (ReferenceObject | ParameterObject)[];
-        requestBody?: ReferenceObject | RequestBodyObject;
-        responses?: ResponsesObject;
-        callbacks?: Record<string, ReferenceObject | CallbackObject>;
-        servers?: ServerObject[];
-    }
+  OpenAPIV3.OperationObject<T>,
+  {
+    parameters?: (ReferenceObject | ParameterObject)[];
+    requestBody?: ReferenceObject | RequestBodyObject;
+    responses?: ResponsesObject;
+    callbacks?: Record<string, ReferenceObject | CallbackObject>;
+    servers?: ServerObject[];
+  }
 > &
-    T;
+T;
 export type ExternalDocumentationObject = OpenAPIV3.ExternalDocumentationObject;
 export interface ParameterObject extends ParameterBaseObject {
-    $id?: string;
-    name: string;
-    in: 'query' | 'header' | 'path' | 'cookie';
+  $id?: string;
+  name: string;
+  in: 'query' | 'header' | 'path' | 'cookie';
 }
 export interface HeaderObject extends ParameterBaseObject {}
 export type ParameterBaseObject = Modify<
-    OpenAPIV3.ParameterBaseObject,
-    {
-        schema?: ReferenceObject | SchemaObject;
-        examples?: {
-            [media: string]: ReferenceObject | ExampleObject;
-        };
-        content?: {
-            [media: string]: MediaTypeObject;
-        };
-    }
+  OpenAPIV3.ParameterBaseObject,
+  {
+    schema?: ReferenceObject | SchemaObject;
+    examples?: {
+      [media: string]: ReferenceObject | ExampleObject;
+    };
+    content?: {
+      [media: string]: MediaTypeObject;
+    };
+  }
 >;
 export type NonArraySchemaObjectType = OpenAPIV3.NonArraySchemaObjectType | 'null';
 export type ArraySchemaObjectType = OpenAPIV3.ArraySchemaObjectType;
@@ -100,106 +98,106 @@ export type ArraySchemaObjectType = OpenAPIV3.ArraySchemaObjectType;
  */
 export type SchemaObject = ArraySchemaObject | NonArraySchemaObject | MixedSchemaObject;
 export interface ArraySchemaObject extends BaseSchemaObject {
-    type: ArraySchemaObjectType;
-    items: ReferenceObject | SchemaObject;
+  type: ArraySchemaObjectType;
+  items: ReferenceObject | SchemaObject;
 }
 export interface NonArraySchemaObject extends BaseSchemaObject {
-    type?: NonArraySchemaObjectType;
+  type?: NonArraySchemaObjectType;
 }
 export interface MixedSchemaObject extends BaseSchemaObject {
-    type?: (ArraySchemaObjectType | NonArraySchemaObjectType)[];
-    items?: ReferenceObject | SchemaObject;
+  type?: (ArraySchemaObjectType | NonArraySchemaObjectType)[];
+  items?: ReferenceObject | SchemaObject;
 }
 export type BaseSchemaObject = Modify<
-    Omit<OpenAPIV3.BaseSchemaObject, 'nullable'>,
-    {
-        $id?: string;
-        $anchor?: string;
-        examples?: OpenAPIV3.BaseSchemaObject['example'][];
-        exclusiveMinimum?: boolean | number;
-        exclusiveMaximum?: boolean | number;
-        contentMediaType?: string;
-        $schema?: string;
-        $vocabulary?: string;
-        additionalProperties?: boolean | ReferenceObject | SchemaObject;
-        properties?: {
-            [name: string]: ReferenceObject | SchemaObject;
-        };
-        allOf?: (ReferenceObject | SchemaObject)[];
-        oneOf?: (ReferenceObject | SchemaObject)[];
-        anyOf?: (ReferenceObject | SchemaObject)[];
-        not?: ReferenceObject | SchemaObject;
-        discriminator?: DiscriminatorObject;
-        externalDocs?: ExternalDocumentationObject;
-        xml?: XMLObject;
-        const?: any;
-    }
+  Omit<OpenAPIV3.BaseSchemaObject, 'nullable'>,
+  {
+    $id?: string;
+    $anchor?: string;
+    examples?: OpenAPIV3.BaseSchemaObject['example'][];
+    exclusiveMinimum?: boolean | number;
+    exclusiveMaximum?: boolean | number;
+    contentMediaType?: string;
+    $schema?: string;
+    $vocabulary?: string;
+    additionalProperties?: boolean | ReferenceObject | SchemaObject;
+    properties?: {
+      [name: string]: ReferenceObject | SchemaObject;
+    };
+    allOf?: (ReferenceObject | SchemaObject)[];
+    oneOf?: (ReferenceObject | SchemaObject)[];
+    anyOf?: (ReferenceObject | SchemaObject)[];
+    not?: ReferenceObject | SchemaObject;
+    discriminator?: DiscriminatorObject;
+    externalDocs?: ExternalDocumentationObject;
+    xml?: XMLObject;
+    const?: any;
+  }
 >;
 export type DiscriminatorObject = OpenAPIV3.DiscriminatorObject;
 export type XMLObject = OpenAPIV3.XMLObject;
 export type ReferenceObject = Modify<
-    OpenAPIV3.ReferenceObject,
-    {
-        summary?: string;
-        description?: string;
-    }
+  OpenAPIV3.ReferenceObject,
+  {
+    summary?: string;
+    description?: string;
+  }
 >;
 export type ExampleObject = OpenAPIV3.ExampleObject;
 export type MediaTypeObject = Modify<
-    OpenAPIV3.MediaTypeObject,
-    {
-        schema?: SchemaObject | ReferenceObject;
-        examples?: Record<string, ReferenceObject | ExampleObject>;
-    }
+  OpenAPIV3.MediaTypeObject,
+  {
+    schema?: SchemaObject | ReferenceObject;
+    examples?: Record<string, ReferenceObject | ExampleObject>;
+  }
 >;
 export type EncodingObject = OpenAPIV3.EncodingObject;
 export type RequestBodyObject = Modify<
-    OpenAPIV3.RequestBodyObject,
-    {
-        $id?: string;
-        content: {
-            [media: string]: MediaTypeObject;
-        };
-    }
+  OpenAPIV3.RequestBodyObject,
+  {
+    $id?: string;
+    content: {
+      [media: string]: MediaTypeObject;
+    };
+  }
 >;
 export type ResponsesObject = Record<string, ReferenceObject | ResponseObject>;
 export type ResponseObject = Modify<
-    OpenAPIV3.ResponseObject,
-    {
-        $id?: string;
-        headers?: {
-            [header: string]: ReferenceObject | HeaderObject;
-        };
-        content?: {
-            [media: string]: MediaTypeObject;
-        };
-        links?: {
-            [link: string]: ReferenceObject | LinkObject;
-        };
-    }
+  OpenAPIV3.ResponseObject,
+  {
+    $id?: string;
+    headers?: {
+      [header: string]: ReferenceObject | HeaderObject;
+    };
+    content?: {
+      [media: string]: MediaTypeObject;
+    };
+    links?: {
+      [link: string]: ReferenceObject | LinkObject;
+    };
+  }
 >;
 export type LinkObject = Modify<
-    OpenAPIV3.LinkObject,
-    {
-        server?: ServerObject;
-    }
+  OpenAPIV3.LinkObject,
+  {
+    server?: ServerObject;
+  }
 >;
 export type CallbackObject = Record<string, PathItemObject | ReferenceObject>;
 export type SecurityRequirementObject = OpenAPIV3.SecurityRequirementObject;
 export type ComponentsObject = Modify<
-    OpenAPIV3.ComponentsObject,
-    {
-        schemas?: Record<string, ReferenceObject | SchemaObject>;
-        responses?: Record<string, ReferenceObject | ResponseObject>;
-        parameters?: Record<string, ReferenceObject | ParameterObject>;
-        examples?: Record<string, ReferenceObject | ExampleObject>;
-        requestBodies?: Record<string, ReferenceObject | RequestBodyObject>;
-        headers?: Record<string, ReferenceObject | HeaderObject>;
-        securitySchemes?: Record<string, ReferenceObject | SecuritySchemeObject>;
-        links?: Record<string, ReferenceObject | LinkObject>;
-        callbacks?: Record<string, ReferenceObject | CallbackObject>;
-        pathItems?: Record<string, ReferenceObject | PathItemObject>;
-    }
+  OpenAPIV3.ComponentsObject,
+  {
+    schemas?: Record<string, ReferenceObject | SchemaObject>;
+    responses?: Record<string, ReferenceObject | ResponseObject>;
+    parameters?: Record<string, ReferenceObject | ParameterObject>;
+    examples?: Record<string, ReferenceObject | ExampleObject>;
+    requestBodies?: Record<string, ReferenceObject | RequestBodyObject>;
+    headers?: Record<string, ReferenceObject | HeaderObject>;
+    securitySchemes?: Record<string, ReferenceObject | SecuritySchemeObject>;
+    links?: Record<string, ReferenceObject | LinkObject>;
+    callbacks?: Record<string, ReferenceObject | CallbackObject>;
+    pathItems?: Record<string, ReferenceObject | PathItemObject>;
+  }
 >;
 export type SecuritySchemeObject = OpenAPIV3.SecuritySchemeObject;
 export type HttpSecurityScheme = OpenAPIV3.HttpSecurityScheme;
