@@ -1,3 +1,3 @@
 export function never(val: never) {
-    //
+  //
 }
