@@ -5,141 +5,138 @@
  * @description This is a sample server Petstore server.  You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).  For this sample, you can use the api key `special-key` to test the authorization filters.
  */
 
-import axios from "axios";
-import type {AxiosRequestConfig, AxiosPromise} from "axios";
-
-
+import type { AxiosPromise, AxiosRequestConfig } from 'axios';
+import axios from 'axios';
 
 // helpers --- start
 type AnyOf<T extends unknown[]> = T extends [infer A, ...infer B] ? A | AnyOf<B> | (A & AnyOf<B>) : never;
 type UnknownObject = Record<string, unknown>;
 type DeepGet<O, K> = K extends [infer P, ...infer R]
-    ? O extends Record<string, unknown> | Array<unknown>
-        ? P extends keyof O
-            ? R['length'] extends 0
-                ? O[P]
-                : DeepGet<NonNullable<O[P]>, R>
-            : never
-        : never
-    : never;
+  ? O extends Record<string, unknown> | Array<unknown>
+    ? P extends keyof O
+      ? R['length'] extends 0
+        ? O[P]
+        : DeepGet<NonNullable<O[P]>, R>
+      : never
+    : never
+  : never;
 // helpers --- end
-    
 
-export type Order = {
+export interface Order {
 /**
  * @format int64
  */
-"id"?:number;
+  id?: number;
+  /**
+   * @format int64
+   */
+  petId?: number;
+  /**
+   * @format int32
+   */
+  quantity?: number;
+  /**
+   * @format date-time
+   */
+  shipDate?: string;
+  /**
+   * @description Order Status
+   */
+  status?: ('placed' | 'approved' | 'delivered');
+  complete?: boolean;
+}
+
+export interface Category {
 /**
  * @format int64
  */
-"petId"?:number;
+  id?: number;
+  name?: string;
+}
+
+export interface User {
+/**
+ * @format int64
+ */
+  id?: number;
+  username?: string;
+  firstName?: string;
+  lastName?: string;
+  email?: string;
+  password?: string;
+  phone?: string;
+  /**
+   * @description User Status
+   * @format int32
+   */
+  userStatus?: number;
+}
+
+export interface Tag {
+/**
+ * @format int64
+ */
+  id?: number;
+  name?: string;
+}
+
+export interface Pet {
+/**
+ * @format int64
+ */
+  id?: number;
+  category?: Category;
+  /**
+   * @example doggie
+   */
+  name: string;
+  photoUrls: Array<string>;
+  tags?: Array<Tag>;
+  /**
+   * @description pet status in the store
+   */
+  status?: ('available' | 'pending' | 'sold');
+}
+
+export interface ApiResponse {
 /**
  * @format int32
  */
-"quantity"?:number;
-/**
- * @format date-time
- */
-"shipDate"?:string;
-/**
- * @description Order Status
- */
-"status"?:("placed"|"approved"|"delivered");
-"complete"?:boolean;
-};
-
-export type Category = {
-/**
- * @format int64
- */
-"id"?:number;
-"name"?:string;
-};
-
-export type User = {
-/**
- * @format int64
- */
-"id"?:number;
-"username"?:string;
-"firstName"?:string;
-"lastName"?:string;
-"email"?:string;
-"password"?:string;
-"phone"?:string;
-/**
- * @description User Status
- * @format int32
- */
-"userStatus"?:number;
-};
-
-export type Tag = {
-/**
- * @format int64
- */
-"id"?:number;
-"name"?:string;
-};
-
-export type Pet = {
-/**
- * @format int64
- */
-"id"?:number;
-"category"?:Category;
-/**
- * @example doggie
- */
-"name":string;
-"photoUrls":Array<string>;
-"tags"?:Array<Tag>;
-/**
- * @description pet status in the store
- */
-"status"?:("available"|"pending"|"sold");
-};
-
-export type ApiResponse = {
-/**
- * @format int32
- */
-"code"?:number;
-"type"?:string;
-"message"?:string;
-};
+  code?: number;
+  type?: string;
+  message?: string;
+}
 
 /**
- * @description 
+ * @description
  * @summary Add a new pet to the store
  * @see pet Everything about your Pets {@link http://swagger.io Find out more}
  * @param data Pet object that needs to be added to the store
  * @param [config] request config
  */
-export async function addPet(data:Pet,config?:AxiosRequestConfig): AxiosPromise<unknown> {
-    return axios({
-        method: "post",
-        url: `/pet`,
-data: data,
-...config
-    });
+export async function addPet(data: Pet, config?: AxiosRequestConfig): AxiosPromise<unknown> {
+  return axios({
+    method: 'post',
+    url: `/pet`,
+    data,
+    ...config,
+  });
 }
 
 /**
- * @description 
+ * @description
  * @summary Update an existing pet
  * @see pet Everything about your Pets {@link http://swagger.io Find out more}
  * @param data Pet object that needs to be added to the store
  * @param [config] request config
  */
-export async function updatePet(data:Pet,config?:AxiosRequestConfig): AxiosPromise<unknown> {
-    return axios({
-        method: "put",
-        url: `/pet`,
-data: data,
-...config
-    });
+export async function updatePet(data: Pet, config?: AxiosRequestConfig): AxiosPromise<unknown> {
+  return axios({
+    method: 'put',
+    url: `/pet`,
+    data,
+    ...config,
+  });
 }
 
 /**
@@ -150,18 +147,18 @@ data: data,
  * @param [config] request config
  * @returns successful operation
  */
-export async function findPetsByStatus(status:Array<
+export async function findPetsByStatus(status: Array<
 /**
  * @default available
  */
-("available"|"pending"|"sold")
->,config?:AxiosRequestConfig): AxiosPromise<Array<Pet>> {
-    return axios({
-        method: "get",
-        url: `/pet/findByStatus`,
-params: {"status": status},
-...config
-    });
+  ('available' | 'pending' | 'sold')
+>, config?: AxiosRequestConfig): AxiosPromise<Array<Pet>> {
+  return axios({
+    method: 'get',
+    url: `/pet/findByStatus`,
+    params: { status },
+    ...config,
+  });
 }
 
 /**
@@ -173,13 +170,13 @@ params: {"status": status},
  * @param [config] request config
  * @returns successful operation
  */
-export async function findPetsByTags(tags:Array<string>,config?:AxiosRequestConfig): AxiosPromise<Array<Pet>> {
-    return axios({
-        method: "get",
-        url: `/pet/findByTags`,
-params: {"tags": tags},
-...config
-    });
+export async function findPetsByTags(tags: Array<string>, config?: AxiosRequestConfig): AxiosPromise<Array<Pet>> {
+  return axios({
+    method: 'get',
+    url: `/pet/findByTags`,
+    params: { tags },
+    ...config,
+  });
 }
 
 /**
@@ -190,59 +187,59 @@ params: {"tags": tags},
  * @param [config] request config
  * @returns successful operation
  */
-export async function getPetById(petId:number,config?:AxiosRequestConfig): AxiosPromise<Pet> {
-    return axios({
-        method: "get",
-        url: `/pet/${petId}`,
-...config
-    });
+export async function getPetById(petId: number, config?: AxiosRequestConfig): AxiosPromise<Pet> {
+  return axios({
+    method: 'get',
+    url: `/pet/${petId}`,
+    ...config,
+  });
 }
 
 /**
- * @description 
+ * @description
  * @summary Updates a pet in the store with form data
  * @see pet Everything about your Pets {@link http://swagger.io Find out more}
  * @param petId ID of pet that needs to be updated
  * @param data request data
  * @param [config] request config
  */
-export async function updatePetWithForm(petId:number,data:{
+export async function updatePetWithForm(petId: number, data: {
 /**
  * @description Updated name of the pet
  */
-"name"?:string;
-/**
- * @description Updated status of the pet
- */
-"status"?:string;
-},config?:AxiosRequestConfig): AxiosPromise<unknown> {
-    return axios({
-        method: "post",
-        url: `/pet/${petId}`,
-data: data,
-...config
-    });
+  name?: string;
+  /**
+   * @description Updated status of the pet
+   */
+  status?: string;
+}, config?: AxiosRequestConfig): AxiosPromise<unknown> {
+  return axios({
+    method: 'post',
+    url: `/pet/${petId}`,
+    data,
+    ...config,
+  });
 }
 
 /**
- * @description 
+ * @description
  * @summary Deletes a pet
  * @see pet Everything about your Pets {@link http://swagger.io Find out more}
  * @param petId Pet id to delete
  * @param [apiKey] request param
  * @param [config] request config
  */
-export async function deletePet(petId:number,apiKey?:string,config?:AxiosRequestConfig): AxiosPromise<unknown> {
-    return axios({
-        method: "delete",
-        url: `/pet/${petId}`,
-headers: {"api_key": apiKey},
-...config
-    });
+export async function deletePet(petId: number, apiKey?: string, config?: AxiosRequestConfig): AxiosPromise<unknown> {
+  return axios({
+    method: 'delete',
+    url: `/pet/${petId}`,
+    headers: { api_key: apiKey },
+    ...config,
+  });
 }
 
 /**
- * @description 
+ * @description
  * @summary uploads an image
  * @see pet Everything about your Pets {@link http://swagger.io Find out more}
  * @param petId ID of pet to update
@@ -250,23 +247,23 @@ headers: {"api_key": apiKey},
  * @param [config] request config
  * @returns successful operation
  */
-export async function uploadFile(petId:number,data:{
+export async function uploadFile(petId: number, data: {
 /**
  * @description Additional data to pass to server
  */
-"additionalMetadata"?:string;
-/**
- * @description file to upload
- * @format binary
- */
-"file"?:Blob;
-},config?:AxiosRequestConfig): AxiosPromise<ApiResponse> {
-    return axios({
-        method: "post",
-        url: `/pet/${petId}/uploadImage`,
-data: data,
-...config
-    });
+  additionalMetadata?: string;
+  /**
+   * @description file to upload
+   * @format binary
+   */
+  file?: Blob;
+}, config?: AxiosRequestConfig): AxiosPromise<ApiResponse> {
+  return axios({
+    method: 'post',
+    url: `/pet/${petId}/uploadImage`,
+    data,
+    ...config,
+  });
 }
 
 /**
@@ -276,34 +273,34 @@ data: data,
  * @param [config] request config
  * @returns successful operation
  */
-export async function getInventory(config?:AxiosRequestConfig): AxiosPromise<{
+export async function getInventory(config?: AxiosRequestConfig): AxiosPromise<{
 /**
  * @format int32
  */
-[key: string]:number;
+  [key: string]: number;
 }> {
-    return axios({
-        method: "get",
-        url: `/store/inventory`,
-...config
-    });
+  return axios({
+    method: 'get',
+    url: `/store/inventory`,
+    ...config,
+  });
 }
 
 /**
- * @description 
+ * @description
  * @summary Place an order for a pet
  * @see store Access to Petstore orders
  * @param data order placed for purchasing the pet
  * @param [config] request config
  * @returns successful operation
  */
-export async function placeOrder(data:Order,config?:AxiosRequestConfig): AxiosPromise<Order> {
-    return axios({
-        method: "post",
-        url: `/store/order`,
-data: data,
-...config
-    });
+export async function placeOrder(data: Order, config?: AxiosRequestConfig): AxiosPromise<Order> {
+  return axios({
+    method: 'post',
+    url: `/store/order`,
+    data,
+    ...config,
+  });
 }
 
 /**
@@ -314,12 +311,12 @@ data: data,
  * @param [config] request config
  * @returns successful operation
  */
-export async function getOrderById(orderId:number,config?:AxiosRequestConfig): AxiosPromise<Order> {
-    return axios({
-        method: "get",
-        url: `/store/order/${orderId}`,
-...config
-    });
+export async function getOrderById(orderId: number, config?: AxiosRequestConfig): AxiosPromise<Order> {
+  return axios({
+    method: 'get',
+    url: `/store/order/${orderId}`,
+    ...config,
+  });
 }
 
 /**
@@ -329,12 +326,12 @@ export async function getOrderById(orderId:number,config?:AxiosRequestConfig): A
  * @param orderId ID of the order that needs to be deleted
  * @param [config] request config
  */
-export async function deleteOrder(orderId:number,config?:AxiosRequestConfig): AxiosPromise<unknown> {
-    return axios({
-        method: "delete",
-        url: `/store/order/${orderId}`,
-...config
-    });
+export async function deleteOrder(orderId: number, config?: AxiosRequestConfig): AxiosPromise<unknown> {
+  return axios({
+    method: 'delete',
+    url: `/store/order/${orderId}`,
+    ...config,
+  });
 }
 
 /**
@@ -344,101 +341,101 @@ export async function deleteOrder(orderId:number,config?:AxiosRequestConfig): Ax
  * @param data Created user object
  * @param [config] request config
  */
-export async function createUser(data:User,config?:AxiosRequestConfig): AxiosPromise<unknown> {
-    return axios({
-        method: "post",
-        url: `/user`,
-data: data,
-...config
-    });
+export async function createUser(data: User, config?: AxiosRequestConfig): AxiosPromise<unknown> {
+  return axios({
+    method: 'post',
+    url: `/user`,
+    data,
+    ...config,
+  });
 }
 
 /**
- * @description 
+ * @description
  * @summary Creates list of users with given input array
  * @see user Operations about user {@link http://swagger.io Find out more about our store}
  * @param data List of user object
  * @param [config] request config
  */
-export async function createUsersWithArrayInput(data:Array<User>,config?:AxiosRequestConfig): AxiosPromise<unknown> {
-    return axios({
-        method: "post",
-        url: `/user/createWithArray`,
-data: data,
-...config
-    });
+export async function createUsersWithArrayInput(data: Array<User>, config?: AxiosRequestConfig): AxiosPromise<unknown> {
+  return axios({
+    method: 'post',
+    url: `/user/createWithArray`,
+    data,
+    ...config,
+  });
 }
 
 /**
- * @description 
+ * @description
  * @summary Creates list of users with given input array
  * @see user Operations about user {@link http://swagger.io Find out more about our store}
  * @param data List of user object
  * @param [config] request config
  */
-export async function createUsersWithListInput(data:Array<User>,config?:AxiosRequestConfig): AxiosPromise<unknown> {
-    return axios({
-        method: "post",
-        url: `/user/createWithList`,
-data: data,
-...config
-    });
+export async function createUsersWithListInput(data: Array<User>, config?: AxiosRequestConfig): AxiosPromise<unknown> {
+  return axios({
+    method: 'post',
+    url: `/user/createWithList`,
+    data,
+    ...config,
+  });
 }
 
 /**
- * @description 
+ * @description
  * @summary Logs user into the system
  * @see user Operations about user {@link http://swagger.io Find out more about our store}
  * @param params request params
  * @param [config] request config
  * @returns successful operation
  */
-export async function loginUser(params:{
+export async function loginUser(params: {
 /**
  * @description The user name for login
  */
-"username":string;
-/**
- * @description The password for login in clear text
- */
-"password":string;
-},config?:AxiosRequestConfig): AxiosPromise<string> {
-    return axios({
-        method: "get",
-        url: `/user/login`,
-params: params,
-...config
-    });
+  username: string;
+  /**
+   * @description The password for login in clear text
+   */
+  password: string;
+}, config?: AxiosRequestConfig): AxiosPromise<string> {
+  return axios({
+    method: 'get',
+    url: `/user/login`,
+    params,
+    ...config,
+  });
 }
 
 /**
- * @description 
+ * @description
  * @summary Logs out current logged in user session
  * @see user Operations about user {@link http://swagger.io Find out more about our store}
  * @param [config] request config
  */
-export async function logoutUser(config?:AxiosRequestConfig): AxiosPromise<unknown> {
-    return axios({
-        method: "get",
-        url: `/user/logout`,
-...config
-    });
+export async function logoutUser(config?: AxiosRequestConfig): AxiosPromise<unknown> {
+  return axios({
+    method: 'get',
+    url: `/user/logout`,
+    ...config,
+  });
 }
 
 /**
- * @description 
+ * @description
  * @summary Get user by user name
  * @see user Operations about user {@link http://swagger.io Find out more about our store}
- * @param username The name that needs to be fetched. Use user1 for testing. 
+ * @param username The name that needs to be fetched. Use user1 for testing.
  * @param [config] request config
  * @returns successful operation
  */
-export async function getUserByName(username:string,config?:AxiosRequestConfig): AxiosPromise<User> {
-    return axios({
-        method: "get",
-        url: `/user/${username}`,
-...config
-    });
+export async function getUserByName(username: string, config?: AxiosRequestConfig): AxiosPromise<User> {
+  return axios({
+    method: 'get',
+    url: `/user/${username}`,
+    ...config,
+  });
 }
 
 /**
@@ -448,12 +445,12 @@ export async function getUserByName(username:string,config?:AxiosRequestConfig):
  * @param username The name that needs to be deleted
  * @param [config] request config
  */
-export async function deleteUser(username:string,config?:AxiosRequestConfig): AxiosPromise<unknown> {
-    return axios({
-        method: "delete",
-        url: `/user/${username}`,
-...config
-    });
+export async function deleteUser(username: string, config?: AxiosRequestConfig): AxiosPromise<unknown> {
+  return axios({
+    method: 'delete',
+    url: `/user/${username}`,
+    ...config,
+  });
 }
 
 /**
@@ -464,11 +461,11 @@ export async function deleteUser(username:string,config?:AxiosRequestConfig): Ax
  * @param data Updated user object
  * @param [config] request config
  */
-export async function updateUser(username:string,data:User,config?:AxiosRequestConfig): AxiosPromise<unknown> {
-    return axios({
-        method: "put",
-        url: `/user/${username}`,
-data: data,
-...config
-    });
+export async function updateUser(username: string, data: User, config?: AxiosRequestConfig): AxiosPromise<unknown> {
+  return axios({
+    method: 'put',
+    url: `/user/${username}`,
+    data,
+    ...config,
+  });
 }
