@@ -6,177 +6,174 @@
  * Swagger at [http://swagger.io](http://swagger.io). In the third iteration of the pet store, we've switched to the design first approach!
  * You can now help us improve the API whether it's by making changes to the definition itself or to the code.
  * That way, with time, we can improve the API in general, and expose some of the new features in OAS3.
- * 
+ *
  * Some useful links:
  * - [The Pet Store repository](https://github.com/swagger-api/swagger-petstore)
  * - [The source API definition for the Pet Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)
  */
 
-import axios from "axios";
-import type {AxiosRequestConfig, AxiosPromise} from "axios";
-
-
+import type { AxiosPromise, AxiosRequestConfig } from 'axios';
+import axios from 'axios';
 
 // helpers --- start
 type AnyOf<T extends unknown[]> = T extends [infer A, ...infer B] ? A | AnyOf<B> | (A & AnyOf<B>) : never;
 type UnknownObject = Record<string, unknown>;
 type DeepGet<O, K> = K extends [infer P, ...infer R]
-    ? O extends Record<string, unknown> | Array<unknown>
-        ? P extends keyof O
-            ? R['length'] extends 0
-                ? O[P]
-                : DeepGet<NonNullable<O[P]>, R>
-            : never
-        : never
-    : never;
+  ? O extends Record<string, unknown> | Array<unknown>
+    ? P extends keyof O
+      ? R['length'] extends 0
+        ? O[P]
+        : DeepGet<NonNullable<O[P]>, R>
+      : never
+    : never
+  : never;
 // helpers --- end
-    
 
-export type Order = {
+export interface Order {
 /**
  * @format int64
  * @example 10
  */
-"id"?:number;
-/**
- * @format int64
- * @example 198772
- */
-"petId"?:number;
-/**
- * @format int32
- * @example 7
- */
-"quantity"?:number;
-/**
- * @format date-time
- */
-"shipDate"?:string;
-/**
- * @description Order Status
- * @example approved
- */
-"status"?:("placed"|"approved"|"delivered");
-"complete"?:boolean;
-};
+  id?: number;
+  /**
+   * @format int64
+   * @example 198772
+   */
+  petId?: number;
+  /**
+   * @format int32
+   * @example 7
+   */
+  quantity?: number;
+  /**
+   * @format date-time
+   */
+  shipDate?: string;
+  /**
+   * @description Order Status
+   * @example approved
+   */
+  status?: ('placed' | 'approved' | 'delivered');
+  complete?: boolean;
+}
 
-export type Customer = {
+export interface Customer {
 /**
  * @format int64
  * @example 100000
  */
-"id"?:number;
-/**
- * @example fehguy
- */
-"username"?:string;
-"address"?:Array<Address>;
-};
+  id?: number;
+  /**
+   * @example fehguy
+   */
+  username?: string;
+  address?: Array<Address>;
+}
 
-export type Address = {
+export interface Address {
 /**
  * @example 437 Lytton
  */
-"street"?:string;
-/**
- * @example Palo Alto
- */
-"city"?:string;
-/**
- * @example CA
- */
-"state"?:string;
-/**
- * @example 94301
- */
-"zip"?:string;
-};
+  street?: string;
+  /**
+   * @example Palo Alto
+   */
+  city?: string;
+  /**
+   * @example CA
+   */
+  state?: string;
+  /**
+   * @example 94301
+   */
+  zip?: string;
+}
 
-export type Category = {
+export interface Category {
 /**
  * @format int64
  * @example 1
  */
-"id"?:number;
-/**
- * @example Dogs
- */
-"name"?:string;
-};
+  id?: number;
+  /**
+   * @example Dogs
+   */
+  name?: string;
+}
 
-export type User = {
+export interface User {
 /**
  * @format int64
  * @example 10
  */
-"id"?:number;
-/**
- * @example theUser
- */
-"username"?:string;
-/**
- * @example John
- */
-"firstName"?:string;
-/**
- * @example James
- */
-"lastName"?:string;
-/**
- * @example john@email.com
- */
-"email"?:string;
-/**
- * @example 12345
- */
-"password"?:string;
-/**
- * @example 12345
- */
-"phone"?:string;
-/**
- * @description User Status
- * @format int32
- * @example 1
- */
-"userStatus"?:number;
-};
+  id?: number;
+  /**
+   * @example theUser
+   */
+  username?: string;
+  /**
+   * @example John
+   */
+  firstName?: string;
+  /**
+   * @example James
+   */
+  lastName?: string;
+  /**
+   * @example john@email.com
+   */
+  email?: string;
+  /**
+   * @example 12345
+   */
+  password?: string;
+  /**
+   * @example 12345
+   */
+  phone?: string;
+  /**
+   * @description User Status
+   * @format int32
+   * @example 1
+   */
+  userStatus?: number;
+}
 
-export type Tag = {
+export interface Tag {
 /**
  * @format int64
  */
-"id"?:number;
-"name"?:string;
-};
+  id?: number;
+  name?: string;
+}
 
-export type Pet = {
+export interface Pet {
 /**
  * @format int64
  * @example 10
  */
-"id"?:number;
-/**
- * @example doggie
- */
-"name":string;
-"category"?:Category;
-"photoUrls":Array<string>;
-"tags"?:Array<Tag>;
-/**
- * @description pet status in the store
- */
-"status"?:("available"|"pending"|"sold");
-};
+  id?: number;
+  /**
+   * @example doggie
+   */
+  name: string;
+  category?: Category;
+  photoUrls: Array<string>;
+  tags?: Array<Tag>;
+  /**
+   * @description pet status in the store
+   */
+  status?: ('available' | 'pending' | 'sold');
+}
 
-export type ApiResponse = {
+export interface ApiResponse {
 /**
  * @format int32
  */
-"code"?:number;
-"type"?:string;
-"message"?:string;
-};
+  code?: number;
+  type?: string;
+  message?: string;
+}
 
 /**
  * @description Add a new pet to the store
@@ -186,13 +183,13 @@ export type ApiResponse = {
  * @param [config] request config
  * @returns Successful operation
  */
-export async function addPet(data:Pet,config?:AxiosRequestConfig): AxiosPromise<Pet> {
-    return axios({
-        method: "post",
-        url: `/pet`,
-data: data,
-...config
-    });
+export async function addPet(data: Pet, config?: AxiosRequestConfig): AxiosPromise<Pet> {
+  return axios({
+    method: 'post',
+    url: `/pet`,
+    data,
+    ...config,
+  });
 }
 
 /**
@@ -203,13 +200,13 @@ data: data,
  * @param [config] request config
  * @returns Successful operation
  */
-export async function updatePet(data:Pet,config?:AxiosRequestConfig): AxiosPromise<Pet> {
-    return axios({
-        method: "put",
-        url: `/pet`,
-data: data,
-...config
-    });
+export async function updatePet(data: Pet, config?: AxiosRequestConfig): AxiosPromise<Pet> {
+  return axios({
+    method: 'put',
+    url: `/pet`,
+    data,
+    ...config,
+  });
 }
 
 /**
@@ -220,13 +217,13 @@ data: data,
  * @param [config] request config
  * @returns successful operation
  */
-export async function findPetsByStatus(status?:("available"|"pending"|"sold"),config?:AxiosRequestConfig): AxiosPromise<Array<Pet>> {
-    return axios({
-        method: "get",
-        url: `/pet/findByStatus`,
-params: {"status": status},
-...config
-    });
+export async function findPetsByStatus(status?:('available' | 'pending' | 'sold'), config?: AxiosRequestConfig): AxiosPromise<Array<Pet>> {
+  return axios({
+    method: 'get',
+    url: `/pet/findByStatus`,
+    params: { status },
+    ...config,
+  });
 }
 
 /**
@@ -237,13 +234,13 @@ params: {"status": status},
  * @param [config] request config
  * @returns successful operation
  */
-export async function findPetsByTags(tags?:Array<string>,config?:AxiosRequestConfig): AxiosPromise<Array<Pet>> {
-    return axios({
-        method: "get",
-        url: `/pet/findByTags`,
-params: {"tags": tags},
-...config
-    });
+export async function findPetsByTags(tags?: Array<string>, config?: AxiosRequestConfig): AxiosPromise<Array<Pet>> {
+  return axios({
+    method: 'get',
+    url: `/pet/findByTags`,
+    params: { tags },
+    ...config,
+  });
 }
 
 /**
@@ -254,59 +251,59 @@ params: {"tags": tags},
  * @param [config] request config
  * @returns successful operation
  */
-export async function getPetById(petId:number,config?:AxiosRequestConfig): AxiosPromise<Pet> {
-    return axios({
-        method: "get",
-        url: `/pet/${petId}`,
-...config
-    });
+export async function getPetById(petId: number, config?: AxiosRequestConfig): AxiosPromise<Pet> {
+  return axios({
+    method: 'get',
+    url: `/pet/${petId}`,
+    ...config,
+  });
 }
 
 /**
- * @description 
+ * @description
  * @summary Updates a pet in the store with form data
  * @see pet Everything about your Pets {@link http://swagger.io Find out more}
  * @param petId ID of pet that needs to be updated
  * @param [params] request params
  * @param [config] request config
  */
-export async function updatePetWithForm(petId:number,params?:{
+export async function updatePetWithForm(petId: number, params?: {
 /**
  * @description Name of pet that needs to be updated
  */
-"name"?:string;
-/**
- * @description Status of pet that needs to be updated
- */
-"status"?:string;
-},config?:AxiosRequestConfig): AxiosPromise<unknown> {
-    return axios({
-        method: "post",
-        url: `/pet/${petId}`,
-params: params,
-...config
-    });
+  name?: string;
+  /**
+   * @description Status of pet that needs to be updated
+   */
+  status?: string;
+}, config?: AxiosRequestConfig): AxiosPromise<unknown> {
+  return axios({
+    method: 'post',
+    url: `/pet/${petId}`,
+    params,
+    ...config,
+  });
 }
 
 /**
- * @description 
+ * @description
  * @summary Deletes a pet
  * @see pet Everything about your Pets {@link http://swagger.io Find out more}
  * @param petId Pet id to delete
  * @param [apiKey] request param
  * @param [config] request config
  */
-export async function deletePet(petId:number,apiKey?:string,config?:AxiosRequestConfig): AxiosPromise<unknown> {
-    return axios({
-        method: "delete",
-        url: `/pet/${petId}`,
-headers: {"api_key": apiKey},
-...config
-    });
+export async function deletePet(petId: number, apiKey?: string, config?: AxiosRequestConfig): AxiosPromise<unknown> {
+  return axios({
+    method: 'delete',
+    url: `/pet/${petId}`,
+    headers: { api_key: apiKey },
+    ...config,
+  });
 }
 
 /**
- * @description 
+ * @description
  * @summary uploads an image
  * @see pet Everything about your Pets {@link http://swagger.io Find out more}
  * @param petId ID of pet to update
@@ -315,14 +312,14 @@ headers: {"api_key": apiKey},
  * @param [config] request config
  * @returns successful operation
  */
-export async function uploadFile(petId:number,data:Blob,additionalMetadata?:string,config?:AxiosRequestConfig): AxiosPromise<ApiResponse> {
-    return axios({
-        method: "post",
-        url: `/pet/${petId}/uploadImage`,
-data: data,
-params: {"additionalMetadata": additionalMetadata},
-...config
-    });
+export async function uploadFile(petId: number, data: Blob, additionalMetadata?: string, config?: AxiosRequestConfig): AxiosPromise<ApiResponse> {
+  return axios({
+    method: 'post',
+    url: `/pet/${petId}/uploadImage`,
+    data,
+    params: { additionalMetadata },
+    ...config,
+  });
 }
 
 /**
@@ -332,17 +329,17 @@ params: {"additionalMetadata": additionalMetadata},
  * @param [config] request config
  * @returns successful operation
  */
-export async function getInventory(config?:AxiosRequestConfig): AxiosPromise<{
+export async function getInventory(config?: AxiosRequestConfig): AxiosPromise<{
 /**
  * @format int32
  */
-[key: string]:number;
+  [key: string]: number;
 }> {
-    return axios({
-        method: "get",
-        url: `/store/inventory`,
-...config
-    });
+  return axios({
+    method: 'get',
+    url: `/store/inventory`,
+    ...config,
+  });
 }
 
 /**
@@ -353,13 +350,13 @@ export async function getInventory(config?:AxiosRequestConfig): AxiosPromise<{
  * @param [config] request config
  * @returns successful operation
  */
-export async function placeOrder(data:Order,config?:AxiosRequestConfig): AxiosPromise<Order> {
-    return axios({
-        method: "post",
-        url: `/store/order`,
-data: data,
-...config
-    });
+export async function placeOrder(data: Order, config?: AxiosRequestConfig): AxiosPromise<Order> {
+  return axios({
+    method: 'post',
+    url: `/store/order`,
+    data,
+    ...config,
+  });
 }
 
 /**
@@ -370,12 +367,12 @@ data: data,
  * @param [config] request config
  * @returns successful operation
  */
-export async function getOrderById(orderId:number,config?:AxiosRequestConfig): AxiosPromise<Order> {
-    return axios({
-        method: "get",
-        url: `/store/order/${orderId}`,
-...config
-    });
+export async function getOrderById(orderId: number, config?: AxiosRequestConfig): AxiosPromise<Order> {
+  return axios({
+    method: 'get',
+    url: `/store/order/${orderId}`,
+    ...config,
+  });
 }
 
 /**
@@ -385,12 +382,12 @@ export async function getOrderById(orderId:number,config?:AxiosRequestConfig): A
  * @param orderId ID of the order that needs to be deleted
  * @param [config] request config
  */
-export async function deleteOrder(orderId:number,config?:AxiosRequestConfig): AxiosPromise<unknown> {
-    return axios({
-        method: "delete",
-        url: `/store/order/${orderId}`,
-...config
-    });
+export async function deleteOrder(orderId: number, config?: AxiosRequestConfig): AxiosPromise<unknown> {
+  return axios({
+    method: 'delete',
+    url: `/store/order/${orderId}`,
+    ...config,
+  });
 }
 
 /**
@@ -400,13 +397,13 @@ export async function deleteOrder(orderId:number,config?:AxiosRequestConfig): Ax
  * @param data Created user object
  * @param [config] request config
  */
-export async function createUser(data:User,config?:AxiosRequestConfig): AxiosPromise<unknown> {
-    return axios({
-        method: "post",
-        url: `/user`,
-data: data,
-...config
-    });
+export async function createUser(data: User, config?: AxiosRequestConfig): AxiosPromise<unknown> {
+  return axios({
+    method: 'post',
+    url: `/user`,
+    data,
+    ...config,
+  });
 }
 
 /**
@@ -417,69 +414,69 @@ data: data,
  * @param [config] request config
  * @returns Successful operation
  */
-export async function createUsersWithListInput(data:Array<User>,config?:AxiosRequestConfig): AxiosPromise<User> {
-    return axios({
-        method: "post",
-        url: `/user/createWithList`,
-data: data,
-...config
-    });
+export async function createUsersWithListInput(data: Array<User>, config?: AxiosRequestConfig): AxiosPromise<User> {
+  return axios({
+    method: 'post',
+    url: `/user/createWithList`,
+    data,
+    ...config,
+  });
 }
 
 /**
- * @description 
+ * @description
  * @summary Logs user into the system
  * @see user Operations about user
  * @param [params] request params
  * @param [config] request config
  * @returns successful operation
  */
-export async function loginUser(params?:{
+export async function loginUser(params?: {
 /**
  * @description The user name for login
  */
-"username"?:string;
-/**
- * @description The password for login in clear text
- */
-"password"?:string;
-},config?:AxiosRequestConfig): AxiosPromise<string> {
-    return axios({
-        method: "get",
-        url: `/user/login`,
-params: params,
-...config
-    });
+  username?: string;
+  /**
+   * @description The password for login in clear text
+   */
+  password?: string;
+}, config?: AxiosRequestConfig): AxiosPromise<string> {
+  return axios({
+    method: 'get',
+    url: `/user/login`,
+    params,
+    ...config,
+  });
 }
 
 /**
- * @description 
+ * @description
  * @summary Logs out current logged in user session
  * @see user Operations about user
  * @param [config] request config
  */
-export async function logoutUser(config?:AxiosRequestConfig): AxiosPromise<unknown> {
-    return axios({
-        method: "get",
-        url: `/user/logout`,
-...config
-    });
+export async function logoutUser(config?: AxiosRequestConfig): AxiosPromise<unknown> {
+  return axios({
+    method: 'get',
+    url: `/user/logout`,
+    ...config,
+  });
 }
 
 /**
- * @description 
+ * @description
  * @summary Get user by user name
  * @see user Operations about user
- * @param username The name that needs to be fetched. Use user1 for testing. 
+ * @param username The name that needs to be fetched. Use user1 for testing.
  * @param [config] request config
  * @returns successful operation
  */
-export async function getUserByName(username:string,config?:AxiosRequestConfig): AxiosPromise<User> {
-    return axios({
-        method: "get",
-        url: `/user/${username}`,
-...config
-    });
+export async function getUserByName(username: string, config?: AxiosRequestConfig): AxiosPromise<User> {
+  return axios({
+    method: 'get',
+    url: `/user/${username}`,
+    ...config,
+  });
 }
 
 /**
@@ -489,12 +486,12 @@ export async function getUserByName(username:string,config?:AxiosRequestConfig):
  * @param username The name that needs to be deleted
  * @param [config] request config
  */
-export async function deleteUser(username:string,config?:AxiosRequestConfig): AxiosPromise<unknown> {
-    return axios({
-        method: "delete",
-        url: `/user/${username}`,
-...config
-    });
+export async function deleteUser(username: string, config?: AxiosRequestConfig): AxiosPromise<unknown> {
+  return axios({
+    method: 'delete',
+    url: `/user/${username}`,
+    ...config,
+  });
 }
 
 /**
@@ -505,11 +502,11 @@ export async function deleteUser(username:string,config?:AxiosRequestConfig): Ax
  * @param data Update an existent user in the store
  * @param [config] request config
  */
-export async function updateUser(username:string,data:User,config?:AxiosRequestConfig): AxiosPromise<unknown> {
-    return axios({
-        method: "put",
-        url: `/user/${username}`,
-data: data,
-...config
-    });
+export async function updateUser(username: string, data: User, config?: AxiosRequestConfig): AxiosPromise<unknown> {
+  return axios({
+    method: 'put',
+    url: `/user/${username}`,
+    data,
+    ...config,
+  });
 }
