@@ -9,93 +9,90 @@
  * @see {@link http://swagger.io Find out more about Swagger}
  */
 
-import axios from "axios";
-import type {AxiosRequestConfig, AxiosPromise} from "axios";
-
-
+import type { AxiosPromise, AxiosRequestConfig } from 'axios';
+import axios from 'axios';
 
 // helpers --- start
 type AnyOf<T extends unknown[]> = T extends [infer A, ...infer B] ? A | AnyOf<B> | (A & AnyOf<B>) : never;
 type UnknownObject = Record<string, unknown>;
 type DeepGet<O, K> = K extends [infer P, ...infer R]
-    ? O extends Record<string, unknown> | Array<unknown>
-        ? P extends keyof O
-            ? R['length'] extends 0
-                ? O[P]
-                : DeepGet<NonNullable<O[P]>, R>
-            : never
-        : never
-    : never;
+  ? O extends Record<string, unknown> | Array<unknown>
+    ? P extends keyof O
+      ? R['length'] extends 0
+        ? O[P]
+        : DeepGet<NonNullable<O[P]>, R>
+      : never
+    : never
+  : never;
 // helpers --- end
-    
 
 /**
  * @description Category
  */
-export type Category = {
+export interface Category {
 /**
  * @format int64
  * @example 1
  */
-"id"?:number;
-/**
- * @example Dogs
- */
-"name"?:string;
-};
+  id?: number;
+  /**
+   * @example Dogs
+   */
+  name?: string;
+}
 
 /**
  * @description Pet
  */
-export type Pet = {
+export interface Pet {
 /**
  * @format int64
  * @example 10
  */
-"id"?:number;
-/**
- * @description Pet Category
- */
-"category"?:unknown;
-/**
- * @example doggie
- */
-"name":string;
-"photoUrls":Array<string>;
-"tags"?:Array<unknown>;
-/**
- * @description pet status in the store
- */
-"status"?:("available"|"pending"|"sold");
-/**
- * @format int32
- * @example 7
- */
-"availableInstances"?:number;
-"petDetailsId"?:DeepGet<PetDetails, ["id"]>;
-"petDetails"?:PetDetails;
-};
+  id?: number;
+  /**
+   * @description Pet Category
+   */
+  category?: unknown;
+  /**
+   * @example doggie
+   */
+  name: string;
+  photoUrls: Array<string>;
+  tags?: Array<unknown>;
+  /**
+   * @description pet status in the store
+   */
+  status?: ('available' | 'pending' | 'sold');
+  /**
+   * @format int32
+   * @example 7
+   */
+  availableInstances?: number;
+  petDetailsId?: DeepGet<PetDetails, ['id']>;
+  petDetails?: PetDetails;
+}
 
-export type PetDetails = {
+export interface PetDetails {
 /**
  * @format int64
  * @example 10
  */
-"id"?:number;
-/**
- * @description PetDetails Category
- */
-"category"?:Category;
-"tag"?:Tag;
-};
+  id?: number;
+  /**
+   * @description PetDetails Category
+   */
+  category?: Category;
+  tag?: Tag;
+}
 
-export type Tag = {
+export interface Tag {
 /**
  * @format int64
  */
-"id"?:number;
-"name"?:string;
-};
+  id?: number;
+  name?: string;
+}
 
 /**
  * @description Update an existing pet by Id
@@ -105,13 +102,13 @@ export type Tag = {
  * @param [config] request config
  * @returns Successful operation
  */
-export async function updatePet(data:Pet,config?:AxiosRequestConfig): AxiosPromise<Pet> {
-    return axios({
-        method: "put",
-        url: `/pet`,
-data: data,
-...config
-    });
+export async function updatePet(data: Pet, config?: AxiosRequestConfig): AxiosPromise<Pet> {
+  return axios({
+    method: 'put',
+    url: `/pet`,
+    data,
+    ...config,
+  });
 }
 
 /**
@@ -122,13 +119,13 @@ data: data,
  * @param [config] request config
  * @returns Successful operation
  */
-export async function addPet(data:Pet,config?:AxiosRequestConfig): AxiosPromise<Pet> {
-    return axios({
-        method: "post",
-        url: `/pet`,
-data: data,
-...config
-    });
+export async function addPet(data: Pet, config?: AxiosRequestConfig): AxiosPromise<Pet> {
+  return axios({
+    method: 'post',
+    url: `/pet`,
+    data,
+    ...config,
+  });
 }
 
 /**
@@ -137,10 +134,10 @@ data: data,
  * @param petId ID of pet that needs to be fetched
  * @param [config] request config
  */
-export async function getPetById(petId:number,config?:AxiosRequestConfig): AxiosPromise<unknown> {
-    return axios({
-        method: "get",
-        url: `/pet/${petId}`,
-...config
-    });
+export async function getPetById(petId: number, config?: AxiosRequestConfig): AxiosPromise<unknown> {
+  return axios({
+    method: 'get',
+    url: `/pet/${petId}`,
+    ...config,
+  });
 }
