@@ -1,76 +1,76 @@
-import fs from 'fs';
-import path from 'path';
-import { configFileNameOrder, resolveConfig, resolveConfigFile, generate } from '../../src/command';
-import { createTempDirname } from '../helpers';
 import type { OpenAPILatest } from '../../src/types/openapi';
+import fs from 'node:fs';
+import path from 'node:path';
+import { configFileNameOrder, generate, resolveConfig, resolveConfigFile } from '../../src/command';
+import { createTempDirname } from '../helpers';
 
-test('resolveConfigFile', async () => {
-    const [cwd, clean] = createTempDirname();
+it('resolveConfigFile', async () => {
+  const [cwd, clean] = createTempDirname();
 
-    expect(resolveConfigFile(cwd)).toBeUndefined();
+  expect(resolveConfigFile(cwd)).toBeUndefined();
 
-    [...configFileNameOrder].reverse().forEach((name, index) => {
-        const file1 = path.join(cwd, name);
-        const file2 = path.join(cwd, configFileNameOrder[configFileNameOrder.length - 1 - index]);
-        fs.writeFileSync(file1, '', 'utf8');
-        expect(resolveConfigFile(cwd)).toBe(file2);
-    });
+  [...configFileNameOrder].reverse().forEach((name, index) => {
+    const file1 = path.join(cwd, name);
+    const file2 = path.join(cwd, configFileNameOrder[configFileNameOrder.length - 1 - index]);
+    fs.writeFileSync(file1, '', 'utf8');
+    expect(resolveConfigFile(cwd)).toBe(file2);
+  });
 
-    clean();
+  clean();
 });
 
-test('resolveConfig', async () => {
-    const [cwd, clean] = createTempDirname();
+it('resolveConfig', async () => {
+  const [cwd, clean] = createTempDirname();
 
-    expect(() => resolveConfig(cwd)).toThrow('配置文件未找到');
+  expect(() => resolveConfig(cwd)).toThrow('配置文件未找到');
 
-    const file = path.join(cwd, configFileNameOrder[0]);
+  const file = path.join(cwd, configFileNameOrder[0]);
 
-    fs.writeFileSync(
-        file,
-        `module.exports = {
+  fs.writeFileSync(
+    file,
+    `module.exports = {
 
 modules: {
     "test": {"document": "test.openapi.json"}
 }
 
 };`,
-        'utf8',
-    );
-    expect(() => resolveConfig(cwd)).not.toThrow();
+    'utf8',
+  );
+  expect(() => resolveConfig(cwd)).not.toThrow();
 
-    clean();
+  clean();
 });
 
-test('run', async () => {
-    const [cwd, clean] = createTempDirname();
-    const file = path.join(cwd, configFileNameOrder[0]);
+it('run', async () => {
+  const [cwd, clean] = createTempDirname();
+  const file = path.join(cwd, configFileNameOrder[0]);
 
-    fs.writeFileSync(
-        path.join(cwd, 'test.openapi.json'),
-        JSON.stringify({
-            info: {
-                title: 'test',
-                version: '1.0.0',
-            },
-            openapi: '3.0.0',
-            paths: {},
-        } as OpenAPILatest.Document),
-    );
-    fs.writeFileSync(
-        file,
-        `module.exports = {
+  fs.writeFileSync(
+    path.join(cwd, 'test.openapi.json'),
+    JSON.stringify({
+      info: {
+        title: 'test',
+        version: '1.0.0',
+      },
+      openapi: '3.0.0',
+      paths: {},
+    } as OpenAPILatest.Document),
+  );
+  fs.writeFileSync(
+    file,
+    `module.exports = {
 
 modules: {
     "test": "test.openapi.json"
 }
 
 };`,
-        'utf8',
-    );
+    'utf8',
+  );
 
-    await generate(cwd);
-    expect(fs.existsSync(path.join(cwd, 'src/apis/test.ts'))).toBe(true);
+  await generate(cwd);
+  expect(fs.existsSync(path.join(cwd, 'src/apis/test.ts'))).toBe(true);
 
-    clean();
+  clean();
 });
