@@ -4,23 +4,23 @@ import { exampleTest } from '../helpers';
 
 // @ref https://github.com/swagger-api/swagger-petstore/tree/v31?tab=readme-ov-file
 
-test('pet-store2.0', () => {
-    exampleTest('2.0', 'pet-store', (document) => {
-        const printer = new Printer(migrate(document));
-        return printer.print();
-    });
+it('pet-store2.0', () => {
+  exampleTest('2.0', 'pet-store', (document) => {
+    const printer = new Printer(migrate(document));
+    return printer.print();
+  });
 });
 
-test('pet-store3.0', () => {
-    exampleTest('3.0', 'pet-store', (document) => {
-        const printer = new Printer(migrate(document));
-        return printer.print();
-    });
+it('pet-store3.0', () => {
+  exampleTest('3.0', 'pet-store', (document) => {
+    const printer = new Printer(migrate(document));
+    return printer.print();
+  });
 });
 
-test('pet-store3.1', () => {
-    exampleTest('3.1', 'pet-store', (document) => {
-        const printer = new Printer(migrate(document));
-        return printer.print();
-    });
+it('pet-store3.1', () => {
+  exampleTest('3.1', 'pet-store', (document) => {
+    const printer = new Printer(migrate(document));
+    return printer.print();
+  });
 });
