@@ -1,30 +1,30 @@
 import { Printer } from '../../src/printer';
 
-test('number', () => {
-    const printer = new Printer({
-        openapi: '3.1.0',
-        info: {
-            title: 'test',
-            version: '1.0.0',
+it('number', () => {
+  const printer = new Printer({
+    openapi: '3.1.0',
+    info: {
+      title: 'test',
+      version: '1.0.0',
+    },
+    components: {
+      schemas: {
+        OrderId: {
+          type: 'integer',
+          format: 'int64',
+          example: 10,
         },
-        components: {
-            schemas: {
-                OrderId: {
-                    type: 'integer',
-                    format: 'int64',
-                    example: 10,
-                },
-            },
-        },
-    });
-    expect(
-        printer.print({
-            hideHeaders: true,
-            hideHelpers: true,
-            hideInfo: true,
-            hideImports: true,
-        }),
-    ).toMatchInlineSnapshot(`
+      },
+    },
+  });
+  expect(
+    printer.print({
+      hideHeaders: true,
+      hideHelpers: true,
+      hideInfo: true,
+      hideImports: true,
+    }),
+  ).toMatchInlineSnapshot(`
       "/**
        * @format int64
        * @example 10
@@ -33,32 +33,32 @@ test('number', () => {
     `);
 });
 
-test('number enum', () => {
-    const printer = new Printer({
-        openapi: '3.1.0',
-        info: {
-            title: 'test',
-            version: '1.0.0',
+it('number enum', () => {
+  const printer = new Printer({
+    openapi: '3.1.0',
+    info: {
+      title: 'test',
+      version: '1.0.0',
+    },
+    components: {
+      schemas: {
+        OrderId: {
+          type: 'integer',
+          format: 'int64',
+          example: 10,
+          enum: [1, 3, 5, 7, 9],
         },
-        components: {
-            schemas: {
-                OrderId: {
-                    type: 'integer',
-                    format: 'int64',
-                    example: 10,
-                    enum: [1, 3, 5, 7, 9],
-                },
-            },
-        },
-    });
-    expect(
-        printer.print({
-            hideHeaders: true,
-            hideHelpers: true,
-            hideInfo: true,
-            hideImports: true,
-        }),
-    ).toMatchInlineSnapshot(`
+      },
+    },
+  });
+  expect(
+    printer.print({
+      hideHeaders: true,
+      hideHelpers: true,
+      hideInfo: true,
+      hideImports: true,
+    }),
+  ).toMatchInlineSnapshot(`
       "/**
        * @format int64
        * @example 10
@@ -67,32 +67,32 @@ test('number enum', () => {
     `);
 });
 
-test('[number, null] enum', () => {
-    const printer = new Printer({
-        openapi: '3.1.0',
-        info: {
-            title: 'test',
-            version: '1.0.0',
+it('[number, null] enum', () => {
+  const printer = new Printer({
+    openapi: '3.1.0',
+    info: {
+      title: 'test',
+      version: '1.0.0',
+    },
+    components: {
+      schemas: {
+        OrderId: {
+          type: ['integer', 'null'],
+          format: 'int64',
+          example: 10,
+          enum: [1, 3, 5, 7, 9],
         },
-        components: {
-            schemas: {
-                OrderId: {
-                    type: ['integer', 'null'],
-                    format: 'int64',
-                    example: 10,
-                    enum: [1, 3, 5, 7, 9],
-                },
-            },
-        },
-    });
-    expect(
-        printer.print({
-            hideHeaders: true,
-            hideHelpers: true,
-            hideInfo: true,
-            hideImports: true,
-        }),
-    ).toMatchInlineSnapshot(`
+      },
+    },
+  });
+  expect(
+    printer.print({
+      hideHeaders: true,
+      hideHelpers: true,
+      hideInfo: true,
+      hideImports: true,
+    }),
+  ).toMatchInlineSnapshot(`
       "/**
        * @format int64
        * @example 10
@@ -101,33 +101,33 @@ test('[number, null] enum', () => {
     `);
 });
 
-test('type[]', () => {
-    const printer = new Printer({
-        openapi: '3.1.0',
-        info: {
-            title: 'test',
-            version: '1.0.0',
+it('type[]', () => {
+  const printer = new Printer({
+    openapi: '3.1.0',
+    info: {
+      title: 'test',
+      version: '1.0.0',
+    },
+    components: {
+      schemas: {
+        Order: {
+          type: ['integer', 'string'],
+          format: 'int64',
+          example: 10,
+          description: 'test1',
+          deprecated: true,
         },
-        components: {
-            schemas: {
-                Order: {
-                    type: ['integer', 'string'],
-                    format: 'int64',
-                    example: 10,
-                    description: 'test1',
-                    deprecated: true,
-                },
-            },
-        },
-    });
-    expect(
-        printer.print({
-            hideHeaders: true,
-            hideHelpers: true,
-            hideInfo: true,
-            hideImports: true,
-        }),
-    ).toMatchInlineSnapshot(`
+      },
+    },
+  });
+  expect(
+    printer.print({
+      hideHeaders: true,
+      hideHelpers: true,
+      hideInfo: true,
+      hideImports: true,
+    }),
+  ).toMatchInlineSnapshot(`
       "/**
        * @description test1
        * @deprecated
@@ -138,50 +138,50 @@ test('type[]', () => {
     `);
 });
 
-test('AllOf primitive', () => {
-    const printer = new Printer({
-        openapi: '3.1.0',
-        info: {
-            title: 'test',
-            version: '1.0.0',
-        },
-        components: {
-            schemas: {
-                User: {
-                    type: ['object', 'null'],
-                    properties: {
-                        username: {
-                            type: 'string',
-                        },
-                    },
-                    required: ['username'],
-                },
-                Order: {
-                    allOf: [
-                        {
-                            type: 'integer',
-                            format: 'int64',
-                            example: 10,
-                            description: 'test1',
-                            deprecated: true,
-                        },
-                        {
-                            $ref: '#/components/schemas/User',
-                            description: 'test2',
-                        },
-                    ],
-                },
+it('allOf primitive', () => {
+  const printer = new Printer({
+    openapi: '3.1.0',
+    info: {
+      title: 'test',
+      version: '1.0.0',
+    },
+    components: {
+      schemas: {
+        User: {
+          type: ['object', 'null'],
+          properties: {
+            username: {
+              type: 'string',
             },
+          },
+          required: ['username'],
         },
-    });
-    expect(
-        printer.print({
-            hideHeaders: true,
-            hideHelpers: true,
-            hideInfo: true,
-            hideImports: true,
-        }),
-    ).toMatchInlineSnapshot(`
+        Order: {
+          allOf: [
+            {
+              type: 'integer',
+              format: 'int64',
+              example: 10,
+              description: 'test1',
+              deprecated: true,
+            },
+            {
+              $ref: '#/components/schemas/User',
+              description: 'test2',
+            },
+          ],
+        },
+      },
+    },
+  });
+  expect(
+    printer.print({
+      hideHeaders: true,
+      hideHelpers: true,
+      hideInfo: true,
+      hideImports: true,
+    }),
+  ).toMatchInlineSnapshot(`
       "export type User = (({
       "username":string;
       })|(null));
@@ -203,36 +203,36 @@ test('AllOf primitive', () => {
     `);
 });
 
-test('explicit array', () => {
-    const printer = new Printer({
-        openapi: '3.1.0',
-        info: {
-            title: 'test',
-            version: '1.0.0',
+it('explicit array', () => {
+  const printer = new Printer({
+    openapi: '3.1.0',
+    info: {
+      title: 'test',
+      version: '1.0.0',
+    },
+    components: {
+      schemas: {
+        Order: {
+          type: 'array',
+          items: {
+            type: 'integer',
+            format: 'int64',
+            example: 10,
+            description: 'test1',
+            deprecated: true,
+          },
         },
-        components: {
-            schemas: {
-                Order: {
-                    type: 'array',
-                    items: {
-                        type: 'integer',
-                        format: 'int64',
-                        example: 10,
-                        description: 'test1',
-                        deprecated: true,
-                    },
-                },
-            },
-        },
-    });
-    expect(
-        printer.print({
-            hideHeaders: true,
-            hideHelpers: true,
-            hideInfo: true,
-            hideImports: true,
-        }),
-    ).toMatchInlineSnapshot(`
+      },
+    },
+  });
+  expect(
+    printer.print({
+      hideHeaders: true,
+      hideHelpers: true,
+      hideInfo: true,
+      hideImports: true,
+    }),
+  ).toMatchInlineSnapshot(`
       "export type Order = Array<
       /**
        * @description test1
@@ -245,36 +245,36 @@ test('explicit array', () => {
     `);
 });
 
-test('generic array', () => {
-    const printer = new Printer({
-        openapi: '3.1.0',
-        info: {
-            title: 'test',
-            version: '1.0.0',
+it('generic array', () => {
+  const printer = new Printer({
+    openapi: '3.1.0',
+    info: {
+      title: 'test',
+      version: '1.0.0',
+    },
+    components: {
+      schemas: {
+        Order: {
+          type: 'array',
+          items: {
+            type: 'integer',
+            format: 'int64',
+            example: 10,
+            description: 'test1',
+            deprecated: true,
+          },
         },
-        components: {
-            schemas: {
-                Order: {
-                    type: 'array',
-                    items: {
-                        type: 'integer',
-                        format: 'int64',
-                        example: 10,
-                        description: 'test1',
-                        deprecated: true,
-                    },
-                },
-            },
-        },
-    });
-    expect(
-        printer.print({
-            hideHeaders: true,
-            hideHelpers: true,
-            hideInfo: true,
-            hideImports: true,
-        }),
-    ).toMatchInlineSnapshot(`
+      },
+    },
+  });
+  expect(
+    printer.print({
+      hideHeaders: true,
+      hideHelpers: true,
+      hideInfo: true,
+      hideImports: true,
+    }),
+  ).toMatchInlineSnapshot(`
       "export type Order = Array<
       /**
        * @description test1
@@ -287,46 +287,46 @@ test('generic array', () => {
     `);
 });
 
-test('explicit object', () => {
-    const printer = new Printer({
-        openapi: '3.1.0',
-        info: {
-            title: 'test',
-            version: '1.0.0',
-        },
-        components: {
-            schemas: {
-                Order: {
-                    type: 'object',
-                    properties: {
-                        aaa: {
-                            type: 'integer',
-                            format: 'int64',
-                            example: 10,
-                            description: 'test1',
-                            deprecated: true,
-                        },
-                        bbb: {
-                            type: 'string',
-                            required: true,
-                        },
-                        ccc: {
-                            type: 'boolean',
-                        },
-                    },
-                    required: ['aaa'],
-                },
+it('explicit object', () => {
+  const printer = new Printer({
+    openapi: '3.1.0',
+    info: {
+      title: 'test',
+      version: '1.0.0',
+    },
+    components: {
+      schemas: {
+        Order: {
+          type: 'object',
+          properties: {
+            aaa: {
+              type: 'integer',
+              format: 'int64',
+              example: 10,
+              description: 'test1',
+              deprecated: true,
             },
+            bbb: {
+              type: 'string',
+              required: true,
+            },
+            ccc: {
+              type: 'boolean',
+            },
+          },
+          required: ['aaa'],
         },
-    });
-    expect(
-        printer.print({
-            hideHeaders: true,
-            hideHelpers: true,
-            hideInfo: true,
-            hideImports: true,
-        }),
-    ).toMatchInlineSnapshot(`
+      },
+    },
+  });
+  expect(
+    printer.print({
+      hideHeaders: true,
+      hideHelpers: true,
+      hideInfo: true,
+      hideImports: true,
+    }),
+  ).toMatchInlineSnapshot(`
       "export type Order = {
       /**
        * @description test1
@@ -341,42 +341,42 @@ test('explicit object', () => {
     `);
 });
 
-test('generic object', () => {
-    const printer = new Printer({
-        openapi: '3.1.0',
-        info: {
-            title: 'test',
-            version: '1.0.0',
-        },
-        components: {
-            schemas: {
-                Pet: {
-                    type: 'object',
-                    required: ['b-b'],
-                    properties: {
-                        aa: {
-                            required: true,
-                            type: 'object',
-                        },
-                        'b-b': {
-                            type: 'object',
-                        },
-                        string: {
-                            type: 'object',
-                        },
-                    },
-                },
+it('generic object', () => {
+  const printer = new Printer({
+    openapi: '3.1.0',
+    info: {
+      title: 'test',
+      version: '1.0.0',
+    },
+    components: {
+      schemas: {
+        Pet: {
+          type: 'object',
+          required: ['b-b'],
+          properties: {
+            'aa': {
+              required: true,
+              type: 'object',
             },
+            'b-b': {
+              type: 'object',
+            },
+            'string': {
+              type: 'object',
+            },
+          },
         },
-    });
-    expect(
-        printer.print({
-            hideHeaders: true,
-            hideHelpers: true,
-            hideInfo: true,
-            hideImports: true,
-        }),
-    ).toMatchInlineSnapshot(`
+      },
+    },
+  });
+  expect(
+    printer.print({
+      hideHeaders: true,
+      hideHelpers: true,
+      hideInfo: true,
+      hideImports: true,
+    }),
+  ).toMatchInlineSnapshot(`
       "export type Pet = {
       "aa":UnknownObject;
       "b-b":UnknownObject;
@@ -385,93 +385,93 @@ test('generic object', () => {
     `);
 });
 
-test('additionalProperties true', () => {
-    const printer = new Printer({
-        openapi: '3.1.0',
-        info: {
-            title: 'test',
-            version: '1.0.0',
+it('additionalProperties true', () => {
+  const printer = new Printer({
+    openapi: '3.1.0',
+    info: {
+      title: 'test',
+      version: '1.0.0',
+    },
+    components: {
+      schemas: {
+        Pet: {
+          type: 'object',
+          additionalProperties: true,
         },
-        components: {
-            schemas: {
-                Pet: {
-                    type: 'object',
-                    additionalProperties: true,
-                },
-            },
-        },
-    });
-    expect(
-        printer.print({
-            hideHeaders: true,
-            hideHelpers: true,
-            hideInfo: true,
-            hideImports: true,
-        }),
-    ).toMatchInlineSnapshot(`"export type Pet = UnknownObject;"`);
+      },
+    },
+  });
+  expect(
+    printer.print({
+      hideHeaders: true,
+      hideHelpers: true,
+      hideInfo: true,
+      hideImports: true,
+    }),
+  ).toMatchInlineSnapshot(`"export type Pet = UnknownObject;"`);
 });
 
-test('additionalProperties false', () => {
-    const printer = new Printer({
-        openapi: '3.1.0',
-        info: {
-            title: 'test',
-            version: '1.0.0',
+it('additionalProperties false', () => {
+  const printer = new Printer({
+    openapi: '3.1.0',
+    info: {
+      title: 'test',
+      version: '1.0.0',
+    },
+    components: {
+      schemas: {
+        Pet: {
+          type: 'object',
+          additionalProperties: false,
         },
-        components: {
-            schemas: {
-                Pet: {
-                    type: 'object',
-                    additionalProperties: false,
-                },
-            },
-        },
-    });
-    expect(
-        printer.print({
-            hideHeaders: true,
-            hideHelpers: true,
-            hideInfo: true,
-            hideImports: true,
-        }),
-    ).toMatchInlineSnapshot(`"export type Pet = {};"`);
+      },
+    },
+  });
+  expect(
+    printer.print({
+      hideHeaders: true,
+      hideHelpers: true,
+      hideInfo: true,
+      hideImports: true,
+    }),
+  ).toMatchInlineSnapshot(`"export type Pet = {};"`);
 });
 
-test('additionalProperties schema type', () => {
-    const printer = new Printer({
-        openapi: '3.1.0',
-        info: {
-            title: 'test',
-            version: '1.0.0',
+it('additionalProperties schema type', () => {
+  const printer = new Printer({
+    openapi: '3.1.0',
+    info: {
+      title: 'test',
+      version: '1.0.0',
+    },
+    components: {
+      schemas: {
+        PetA: {
+          type: 'object',
+          additionalProperties: {
+            type: 'string',
+            enum: ['a', 'b'],
+          },
         },
-        components: {
-            schemas: {
-                PetA: {
-                    type: 'object',
-                    additionalProperties: {
-                        type: 'string',
-                        enum: ['a', 'b'],
-                    },
-                },
-                PetB: {
-                    type: 'object',
-                    additionalProperties: {
-                        type: 'string',
-                        required: true,
-                        enum: ['a', 'b'],
-                    },
-                },
-            },
+        PetB: {
+          type: 'object',
+          additionalProperties: {
+            type: 'string',
+            required: true,
+            enum: ['a', 'b'],
+          },
         },
-    });
-    expect(
-        printer.print({
-            hideHeaders: true,
-            hideHelpers: true,
-            hideInfo: true,
-            hideImports: true,
-        }),
-    ).toMatchInlineSnapshot(`
+      },
+    },
+  });
+  expect(
+    printer.print({
+      hideHeaders: true,
+      hideHelpers: true,
+      hideInfo: true,
+      hideImports: true,
+    }),
+  ).toMatchInlineSnapshot(`
       "export type PetA = {
       [key: string]:("a"|"b");
       };
@@ -482,46 +482,46 @@ test('additionalProperties schema type', () => {
     `);
 });
 
-test('additionalProperties schema ref', () => {
-    const printer = new Printer({
-        openapi: '3.1.0',
-        info: {
-            title: 'test',
-            version: '1.0.0',
+it('additionalProperties schema ref', () => {
+  const printer = new Printer({
+    openapi: '3.1.0',
+    info: {
+      title: 'test',
+      version: '1.0.0',
+    },
+    components: {
+      schemas: {
+        PetA: {
+          type: 'object',
+          additionalProperties: {
+            $ref: '#/components/schemas/Pet1',
+          },
         },
-        components: {
-            schemas: {
-                PetA: {
-                    type: 'object',
-                    additionalProperties: {
-                        $ref: '#/components/schemas/Pet1',
-                    },
-                },
-                PetB: {
-                    additionalProperties: {
-                        $ref: '#/components/schemas/Pet2',
-                    },
-                },
-                Pet1: {
-                    type: 'string',
-                    enum: ['a', 'b'],
-                },
-                Pet2: {
-                    type: 'string',
-                    required: true,
-                    enum: ['a', 'b'],
-                },
-            },
+        PetB: {
+          additionalProperties: {
+            $ref: '#/components/schemas/Pet2',
+          },
         },
-    });
-    expect(
-        printer.print({
-            hideHeaders: true,
-            hideHelpers: true,
-            hideInfo: true,
-            hideImports: true,
-        }),
-    ).toMatchInlineSnapshot(`
+        Pet1: {
+          type: 'string',
+          enum: ['a', 'b'],
+        },
+        Pet2: {
+          type: 'string',
+          required: true,
+          enum: ['a', 'b'],
+        },
+      },
+    },
+  });
+  expect(
+    printer.print({
+      hideHeaders: true,
+      hideHelpers: true,
+      hideInfo: true,
+      hideImports: true,
+    }),
+  ).toMatchInlineSnapshot(`
       "export type PetA = {
       [key: string]:Pet1;
       };
