@@ -1,52 +1,52 @@
 import { Printer } from '../../src/printer';
 
-test('unique vars', () => {
-    const printer = new Printer({
-        openapi: '3.1.0',
-        info: {
-            title: 'test',
-            version: '1.0.0',
-        },
-        paths: {
-            '/test': {
-                get: {
-                    responses: {
-                        200: {
-                            description: 'success',
-                            content: {
-                                'application/json': {
-                                    schema: {
-                                        $ref: '#/components/schemas/Test_aa_',
-                                    },
-                                },
-                            },
-                        },
-                    },
+it('unique vars', () => {
+  const printer = new Printer({
+    openapi: '3.1.0',
+    info: {
+      title: 'test',
+      version: '1.0.0',
+    },
+    paths: {
+      '/test': {
+        get: {
+          responses: {
+            200: {
+              description: 'success',
+              content: {
+                'application/json': {
+                  schema: {
+                    $ref: '#/components/schemas/Test_aa_',
+                  },
                 },
+              },
             },
+          },
         },
-        components: {
-            schemas: {
-                Test_aa_: {
-                    type: 'object',
-                    properties: {
-                        name: {
-                            type: 'string',
-                        },
-                    },
-                },
+      },
+    },
+    components: {
+      schemas: {
+        Test_aa_: {
+          type: 'object',
+          properties: {
+            name: {
+              type: 'string',
             },
+          },
         },
-    });
+      },
+    },
+  });
 
-    expect(
-        printer.print({
-            hideHeaders: true,
-            hideHelpers: true,
-            hideInfo: true,
-            hideImports: true,
-        }),
-    ).toMatchInlineSnapshot(`
+  expect(
+    printer.print({
+      hideHeaders: true,
+      hideHelpers: true,
+      hideInfo: true,
+      hideImports: true,
+    }),
+  ).toMatchInlineSnapshot(`
       "export type TestAa = {
       "name"?:string;
       };
@@ -65,85 +65,85 @@ test('unique vars', () => {
     `);
 });
 
-test('unique types', () => {
-    const printer = new Printer({
-        openapi: '3.1.0',
-        info: {
-            title: 'test',
-            version: '1.0.0',
-        },
-        paths: {
-            '/test/{axios}': {
-                get: {
-                    parameters: [
-                        {
-                            name: 'axios',
-                            in: 'path',
-                            schema: {
-                                type: 'string',
-                            },
-                        },
-                        {
-                            name: 'UnknownObject',
-                            in: 'query',
-                            schema: {
-                                $ref: '#/components/schemas/UnknownObject',
-                            },
-                        },
-                    ],
-                    responses: {
-                        200: {
-                            description: 'success',
-                            content: {
-                                'application/json': {
-                                    schema: {
-                                        type: 'object',
-                                        properties: {
-                                            AxiosPromise: {
-                                                $ref: '#/components/schemas/AxiosPromise',
-                                            },
-                                            UnknownObject: {
-                                                $ref: '#/components/schemas/UnknownObject',
-                                            },
-                                        },
-                                    },
-                                },
-                            },
-                        },
-                    },
-                },
+it('unique types', () => {
+  const printer = new Printer({
+    openapi: '3.1.0',
+    info: {
+      title: 'test',
+      version: '1.0.0',
+    },
+    paths: {
+      '/test/{axios}': {
+        get: {
+          parameters: [
+            {
+              name: 'axios',
+              in: 'path',
+              schema: {
+                type: 'string',
+              },
             },
-        },
-        components: {
-            schemas: {
-                AxiosPromise: {
+            {
+              name: 'UnknownObject',
+              in: 'query',
+              schema: {
+                $ref: '#/components/schemas/UnknownObject',
+              },
+            },
+          ],
+          responses: {
+            200: {
+              description: 'success',
+              content: {
+                'application/json': {
+                  schema: {
                     type: 'object',
                     properties: {
-                        name: {
-                            type: 'string',
-                        },
+                      AxiosPromise: {
+                        $ref: '#/components/schemas/AxiosPromise',
+                      },
+                      UnknownObject: {
+                        $ref: '#/components/schemas/UnknownObject',
+                      },
                     },
+                  },
                 },
-                UnknownObject: {
-                    type: 'object',
-                    properties: {
-                        name: {
-                            type: 'number',
-                        },
-                    },
-                },
+              },
             },
+          },
         },
-    });
+      },
+    },
+    components: {
+      schemas: {
+        AxiosPromise: {
+          type: 'object',
+          properties: {
+            name: {
+              type: 'string',
+            },
+          },
+        },
+        UnknownObject: {
+          type: 'object',
+          properties: {
+            name: {
+              type: 'number',
+            },
+          },
+        },
+      },
+    },
+  });
 
-    expect(
-        printer.print({
-            hideHeaders: true,
-            hideHelpers: true,
-            hideInfo: true,
-            hideImports: true,
-        }),
-    ).toMatchInlineSnapshot(`
+  expect(
+    printer.print({
+      hideHeaders: true,
+      hideHelpers: true,
+      hideInfo: true,
+      hideImports: true,
+    }),
+  ).toMatchInlineSnapshot(`
       "export type AxiosPromise_2 = {
       "name"?:string;
       };
