@@ -1,28 +1,28 @@
 import { Named } from '../../src/printer/Named';
 import { Schemata } from '../../src/printer/Schemata';
 
-test('types', () => {
-    const named = new Named();
-    const schemata = new Schemata(named);
-    const type = schemata.toString({
-        type: 'object',
-        properties: {
-            t0: {
-                type: 'string',
-            },
-            t1: {
-                type: [],
-            },
-            t2: {
-                type: ['string'],
-            },
-            t3: {
-                type: ['string', 'number'],
-            },
-        },
-    });
+it('types', () => {
+  const named = new Named();
+  const schemata = new Schemata(named);
+  const type = schemata.toString({
+    type: 'object',
+    properties: {
+      t0: {
+        type: 'string',
+      },
+      t1: {
+        type: [],
+      },
+      t2: {
+        type: ['string'],
+      },
+      t3: {
+        type: ['string', 'number'],
+      },
+    },
+  });
     // console.log(type);
-    expect(type).toMatchInlineSnapshot(`
+  expect(type).toMatchInlineSnapshot(`
       "{
       "t0"?:string;
       "t1"?:unknown;
@@ -32,19 +32,19 @@ test('types', () => {
     `);
 });
 
-test('UnknownObject', () => {
-    const named = new Named();
-    const schemata = new Schemata(named);
-    const type = schemata.toString({
+it('unknownObject', () => {
+  const named = new Named();
+  const schemata = new Schemata(named);
+  const type = schemata.toString({
+    type: 'object',
+    properties: {
+      t0: {
         type: 'object',
-        properties: {
-            t0: {
-                type: 'object',
-            },
-        },
-    });
+      },
+    },
+  });
     // console.log(type);
-    expect(type).toMatchInlineSnapshot(`
+  expect(type).toMatchInlineSnapshot(`
       "{
       "t0"?:UnknownObject;
       }"
